osif/requirements.txt
sploitkit>=0.5.5
terminaltables
vt-py
python-dotenv
sublist3r
dnspython
win_unicode_console
censys
malwarebazaar
python-Wappalyzer
shodan
requests_oauthlib
jellyfish